backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
cryptography
pyotp
argon2-cffi